requirements.txt
apispec==4.0.0
attrs==20.3.0
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
# flasgger==0.9.5
Flask==1.1.2
Flask-Caching==1.9.0
# flask-marshmallow==0.14.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
jsonschema==3.2.0
MarkupSafe==1.1.1
# marshmallow==3.9.1
mistune==0.8.4
psycopg2==2.8.6
pycodestyle==2.6.0
pyrsistent==0.17.3
python-decouple==3.3
python-dotenv==0.15.0
PyYAML==5.3.1
requests==2.25.0
roman==3.3
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
